fix: 修復因TypeScript所產生的警告，把Labels的Context跟Provider分開來
