+++ taskify-frontend/src/components/AddTask.tsx
@@ -7,7 +7,7 @@ import { addTask } from "@/api/tasks";
 import { notifications } from "@mantine/notifications";
 import { BaseDataRes, BaseTaskRes, ColumnResType } from "@/types/column";
 import { calculateDataIndex } from "@/utils";
-import { v4 as uuidv4 } from "uuid";
+import { v4 as uuidV4 } from "uuid";
 
 type Props = {
   isAddingTask: boolean;
@@ -34,7 +34,7 @@ function AddTask({ isAddingTask, toggleAddingTask, column }: Props) {
     onMutate: async (variables) => {
       await queryClient.cancelQueries({ queryKey: ["tasks"] });
       const optimisticTask: BaseTaskRes = {
-        id: uuidv4(),
+        id: uuidV4(),
         name: variables.name,
         dataIndex: variables.dataIndex,
         labels: [],
@@ -49,7 +49,7 @@ function AddTask({ isAddingTask, toggleAddingTask, column }: Props) {
       });
       return { optimisticTask };
     },
-    onSuccess: (resData: BaseTaskRes, variables, context) => {
+    onSuccess: (resData: BaseTaskRes, _variables, context) => {
       queryClient.setQueryData(["tasks"], (oldData: BaseDataRes) => {
         notifications.show({
           icon: <IconMoodCheck />,
@@ -67,7 +67,7 @@ function AddTask({ isAddingTask, toggleAddingTask, column }: Props) {
         };
       });
     },
-    onError: (err, variables, context) => {
+    onError: (_err, _variables, context) => {
       queryClient.setQueryData(["tasks"], (oldData: BaseDataRes) => {
         return {
           ...oldData,
